Move desired state to new component
